Conflicts:
	hibernate-core/src/main/java/org/hibernate/dialect/HSQLDialect.java

+++ hibernate-core/src/main/java/org/hibernate/tuple/component/ComponentMetamodel.java
@@ -23,6 +23,7 @@
  *
  */
 package org.hibernate.tuple.component;
+
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -107,7 +108,7 @@ public class ComponentMetamodel implements Serializable {
 		if ( index == null ) {
 			throw new HibernateException( "component does not contain such a property [" + propertyName + "]" );
 		}
-		return index.intValue();
+		return index;
 	}
 
 	public StandardProperty getProperty(String propertyName) {
